requirements.txt
asgiref==3.9.1                  # requerido por Django
certifi==2025.8.3               # requests/urllib3
charset-normalizer==3.4.3       # requests/urllib3
Django==5.1.1                   # framework principal
django-autocomplete-light==3.11.0  # autocompletado en formularios
django-widget-tweaks==1.5.0     # personalización de templates
idna==3.10                      # requests
psycopg2-binary==2.9.10         # conexión a PostgreSQL
python-decouple==3.8            # manejo de variables de entorno (.env)
requests==2.32.5                # llamadas HTTP
sqlparse==0.5.3                 # requerido por Django
urllib3==2.5.0                  # requests
websocket-client==1.8.0         # conexiones WebSocket
whitenoise==6.10.0              # servir archivos estáticos
django-csp>=4.0,<5.0            #imprimir facturas